Refactored channel code to reduce duplicate code.
